--- readme.md
@@ -1,50 +1,0 @@
-# Работа с maven
-## Что понадобится для выполнения
-* Аккаунт на [Github](https://github.com)
-* Установленный [maven](https://maven.apache.org)
-* Настроенная JDK версии 8+
-* Любая удобная IDE ([Intellij Idea](https://www.jetbrains.com/idea/), [Netbeans](https://netbeans.org/), [Eclipse](https://www.eclipse.org/))
-## Предусловия
-Для решения необходимо будет сделать fork репозитория. И все действия будут производится с ним.
-## Задание
-### Вариант 1
-Необходимо разработать новый плагин, который будет выполнять какую-то полезную функциональность. Через вызов команды хелп должно становится понятным, что он делает. Не требуется писать сложный плагин, но это не должно быть HelloWorld.
-
-Обязательные требования:
-* плагин должен принимать не менее 2 входных параметров
-* должна быть дефолтная конфигурация
-
-
-Примеры плагинов:
-* сделать плагин, который проверяет, что в папке не меньше определенного количества файлов
-* сделать плагин, который проверяет, что нет файлов больше определенного размера
-* сделать плагин, который отправляет email сообщение после успешной сборки
-и т.д.
-### Вариант 2
-Написать небольшое приложение, используя maven
-
-
-Обязательные требования:
-* должны быть задействованы checkstyle plugin, surefire plugin, resource plugin, release plugin
-* должно быть не менее 2 профилей
-* должно быть не менее 3 внешних зависимостей
-* должно быть не менее 10 Java классов
-* должны присутствовать тесты
-
-Написать можно любое приложение, главное, чтобы выполняло требования.
-
-Код должен быть хорошего качества
-## Решения
-Решения по задачам должны быть оформлены в виде PullRequest. В заголовке должно быть ФИО.
-## Сроки
-Срока будет 2 - мягкий и жесткий
-
-
-**Мягкий** - 26 декабря
- 
-**Жесткий** - 11 января
-
-**После мягкого срока будет снижена суммарная оценка. А после жёсткого результат уже не будет засчитан.**
-
-## P.S.
-Ревью чужих PullRequest будет поощряться
