+++ readme_rogozhinski.md
@@ -1,0 +1,24 @@
+## Домашнее задание по maven
+
+	В качестве направления домашнего задания выбран второй вариант - создание приложения.
+
+## Описание приложения
+
+	Приложение предназначено для выполнения расчетов на основе некоторых из основных физических формул, среди которых:
+	- второй закон Ньютона;
+	- уравнение гидростатического давления;
+	- закон Ома;
+	- выражение для энергии конденсатора.
+
+	Приложение работает в консольном режиме и управляется с помощью численных значений, вводимых пользователем с клавиатуры. Случаи некорректного ввода обрабатываются.
+
+## Сложности 
+
+	1. На данный момент возникла сложность с запуском jar-файла. Несмотря на удачную компиляцию проекта и сборку .jar, запуск последнего приводит к следующей ошибке:
+
+	"Error: Could not find or load main class com.Application"
+	
+	При этом в POM'e все ссылки на основной класс приложения заменены на актуальные. Попытки запуска вне IDE (с консоли ОС) не приводят к успеху
+
+	2. Ввиду отсутствия опыта в подготовке тестов, они на данный момент находятся все еще на стадии разработки. В частности, трудности вызвала потребность в эмуляции ввода с клавиатуры.
+
