--- src/Application.java
@@ -1,11 +1,0 @@
-import app_interface.MainInterface;
-
-public class Application {
-
-    public static void main(String[] args) {
-
-        MainInterface app_interface = new MainInterface ();
-        app_interface.choosing_physics_branch();
-
-    }
-}
