+++ src/main/java/com/Application.java
@@ -1,0 +1,13 @@
+package main.java.com;
+
+import main.java.com.app_interface.MainInterface;
+
+public class Application {
+
+    public static void main(String[] args) {
+
+        MainInterface app_interface = new MainInterface();
+        app_interface.choosing_physics_branch();
+
+    }
+}
