+++ src/main/java/com/mechanics/InertiaLaw.java
@@ -1,0 +1,20 @@
+package main.java.com.mechanics;
+
+public class InertiaLaw {
+
+    double force;
+    double mass;
+    double acceleration;
+
+    public InertiaLaw(double f, double m, double a){
+        force = f;
+        mass = m;
+        acceleration = a;
+    }
+
+    public double force_computation () {return mass * acceleration;};
+
+    public double mass_computation () {return force / acceleration;};
+
+    public double acceleration_computation () {return force / mass;};
+}
